Add a title (#38)
## Description
<!---
Your title should be concise. And this section should be more thorough.
-->

## Related Issue
<!---
Starting with an open issue is a great way to make sure your requested
changes
are understood and follow the intended purpose of the repo. It would not
be fun
to write up a whole new feature and find out that it's a hard no after
writing a
bunch of code.

- If suggesting a new feature or change, discuss it in an issue first
- If fixing a bug, there should be an issue describing it with steps to
reproduce
 
 Please link to the issue here:
 -->

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
<!--- 
Good testing gives reviewers confidence!

Please describe in detail how you tested your changes.
Include details of your testing environment, and the tests you ran to
see how your change affects other areas of the code, etc.
-->

## Screenshots (if appropriate):

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.